finished up the xsrf protection

+++ users.py
@@ -6,24 +6,33 @@ cgi = database.cgi
 
 class MainHandler(database.webapp2.RequestHandler):
   def get(self):
-    li = database.db.GqlQuery("SELECT * FROM LoginInformation WHERE user_id = :1", database.users.get_current_user().user_id()).get()
-    database.render_template(self, '/users/index.html', {'li': li})
+    user = database.users.get_current_user()
+    if user:
+      li = database.get_current_li()
+      li.create_xsrf_token()
+      database.render_template(self, '/users/index.html', {'li': li})
+    else:
+      self.redirect('/')
 
 class RegisterHandler(database.webapp2.RequestHandler):
   def get(self):
     user = database.users.get_current_user()
-    li = database.db.GqlQuery("SELECT * FROM LoginInformation WHERE user_id = :1", user.user_id())
-    if li.count() == 1:
-      self.redirect('/')
+    if user:
+      li = database.db.GqlQuery("SELECT * FROM LoginInformation WHERE user_id = :1", user.user_id())
+      if li.count() == 1:
+        self.redirect('/')
+      else:
+        li.create_xsrf_token()
+        database.render_template(self, '/users/register_user.html', {})
     else:
-      database.render_template(self, '/users/register_user.html', {})
+      self.redirect('/')
       
 class SaveLIHandler(database.webapp2.RequestHandler):
   def post(self):
     user = database.users.get_current_user()
     li = database.db.GqlQuery("SELECT * FROM LoginInformation WHERE user_id = :1", user.user_id())
     #check for duplicates
-    if li.count() == 0:
+    if user and li.count() == 0 and li.verify_xsrf_token(self):
       li = database.LoginInformation()
       li.first_name = cgi.escape(self.request.get('first_name'))
       li.last_name = cgi.escape(self.request.get('last_name'))
@@ -41,14 +50,17 @@ class SaveLIHandler(database.webapp2.RequestHandler):
 class DeleteHandler(database.webapp2.RequestHandler):
   def get(self):
     user = database.users.get_current_user()
-    #delete all the items
-    items = database.db.GqlQuery("SELECT * FROM Item WHERE created_by_id = :1", user.user_id())
-    for item in items:
-      database.db.delete(item)
-    #delete the li
-    li = database.db.GqlQuery("SELECT * FROM LoginInformation WHERE user_id = :1", user.user_id()).get()
-    database.db.delete(li)
-    self.redirect(database.users.create_logout_url('/'))
+    if user and database.get_current_li().verify_xsrf_token(self):
+      #delete all the items
+      items = database.db.GqlQuery("SELECT * FROM Item WHERE created_by_id = :1", user.user_id())
+      for item in items:
+        database.db.delete(item)
+      #delete the li
+      li = database.get_current_li()
+      database.db.delete(li)
+      self.redirect(database.users.create_logout_url('/'))
+    else:
+      self.redirect('/')
     
     
 app = database.webapp2.WSGIApplication([('/users/', MainHandler), ('/users/verify_user/', RegisterHandler),
